Вынесена общая логика обновления
